requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector==2.2.9
numpy==2.2.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
